Make draw functions take floats instead of doubles

+++ projects/server/apps/DarkWorld/src/common/graphics/draw.cpp
@@ -9,12 +9,12 @@ namespace dw {
                                  std::to_string(b));
     }
 
-    void draw_image(std::shared_ptr<net_session> net_session, std::string_view image_name, double x, double y, double w,
-                    double h) {
+    void draw_image(std::shared_ptr<net_session> net_session, std::string_view image_name, float x, float y, float w,
+                    float h) {
         net_session->add_message(std::string("draw_image;") + image_name.data() + ";" + std::to_string(x) + ";" +
                                  std::to_string(y) + ";" + std::to_string(w) + ";" + std::to_string(h));
     }
-    void draw_text(std::shared_ptr<net_session> net_session, std::string_view text, double x, double y, int r, int g,
+    void draw_text(std::shared_ptr<net_session> net_session, std::string_view text, float x, float y, int r, int g,
                    int b) {
         net_session->add_message(std::string("draw_text;") + text.data() + ";" + std::to_string(x) + ";" +
                                  std::to_string(y) + ";" + std::to_string(r) + ";" + std::to_string(g) + ";" +
